get(User)

+++ Feed/src/Notification/Main.java
@@ -7,7 +7,7 @@ public class Main {
 
 	public static void main(String[] args) {
 
-		Notify nt = new Notify("oila", 0, 0, TYPE_NOTIFY.FRIENDSHIP);
+		Notify nt = new Notify("oila", (long) 0, (long) 0, TYPE_NOTIFY.FRIENDSHIP);
 		User u = new User("matehus");
 		// // Post p = new Post("--", "==", 10, "nada");
 		Data.getDao().save(u);
@@ -16,7 +16,7 @@ public class Main {
 		// Data.getDao().save(u);
 		// Data.getDao().save(p);
 		Data.load();
-//		Notification.birth_day(Data.getUsers().get(0));
+		// Notification.birth_day(Data.getUsers().get(0));
 		//
 	}
 }
