+++ Feed/src/Object/User.java
@@ -16,13 +16,13 @@ public class User {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	@Column
 	private String name;
 	@Column
 	private Calendar birth;
 	@Column
-	private ArrayList<Integer> friends;
+	private ArrayList<Long> friends;
 
 	public User() {
 		super();
@@ -50,11 +50,11 @@ public class User {
 		this.birth = birth;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -66,11 +66,11 @@ public class User {
 		this.name = name;
 	}
 
-	public ArrayList<Integer> getFriends() {
+	public ArrayList<Long> getFriends() {
 		return friends;
 	}
 
-	public void setFriends(ArrayList<Integer> friends) {
+	public void setFriends(ArrayList<Long> friends) {
 		this.friends = friends;
 	}
